feat: 목업과 도메인 개발 환경 분리
